Storybook: Brush - Split into simple API story and additional example
