+++ storybook/stories/API/cartesian/Brush.stories.tsx
@@ -1,15 +1,9 @@
 import React from 'react';
-// eslint-disable-next-line import/no-extraneous-dependencies
 import { fireEvent } from '@storybook/testing-library';
 import { Args } from '@storybook/react';
-import { numberData, dateData } from '../../data';
-import { Surface, Brush, ResponsiveContainer, LineChart, Line } from '../../../../src';
+import { pageData } from '../../data';
+import { Brush, ResponsiveContainer, ComposedChart, Line } from '../../../../src';
 import { getStoryArgsFromArgsTypesObject } from '../props/utils';
-
-interface BrushStartEndIndex {
-  startIndex?: number;
-  endIndex?: number;
-}
 
 const GeneralProps: Args = {
   dataKey: {
@@ -42,7 +36,7 @@ const GeneralProps: Args = {
     defaultValue: 5,
   },
   gap: {
-    description: `The data with gap of refreshing chart. If the option is not set, 
+    description: `The data with gap of refreshing chart. If the option is not set,
     the chart will be refreshed every time.`,
     table: { type: { summary: 'number' }, category: 'General' },
     defaultValue: 1,
@@ -52,7 +46,7 @@ const GeneralProps: Args = {
     table: { type: { summary: 'number' }, category: 'General' },
   },
   endIndex: {
-    description: `The default end index of brush. If the option is not set, 
+    description: `The default end index of brush. If the option is not set,
     the end index will be calculated by the length of data`,
     table: { type: { summary: 'number' }, category: 'General' },
   },
@@ -72,121 +66,26 @@ const GeneralProps: Args = {
 };
 
 export default {
+  component: Brush,
   argTypes: {
     ...GeneralProps,
   },
-  component: Brush,
 };
 
-const [surfaceWidth, surfaceHeight] = [600, 500];
-
-export const Simple = {
-  render: (args: Record<string, any>) => {
-    const [simple, setSimple] = React.useState<BrushStartEndIndex>({
-      startIndex: 0,
-      endIndex: args.data.length - 1,
-    });
-
-    const [gap, setGap] = React.useState<BrushStartEndIndex>({
-      startIndex: 0,
-      endIndex: args.data.length - 1,
-    });
-
-    const handleChange = (res: BrushStartEndIndex) => {
-      setSimple(res);
-    };
-
-    const handleGapChange = (res: BrushStartEndIndex) => {
-      setGap(res);
-    };
-
-    // eslint-disable-next-line no-shadow
-    const renderTraveller = (props: { x: number; y: number; width: number; height: number }) => {
-      const { x, y, width, height } = props;
-
-      return (
-        <path
-          d={`M${x + width / 2},${y}L${x + width},${y + height / 2}L${x + width / 2},${y + height}L${x},${
-            y + height / 2
-          }Z`}
-          fill="red"
-          stroke="none"
-        />
-      );
-    };
-
-    return (
-      <ResponsiveContainer width="100%" height={surfaceHeight}>
-        <div style={{ margin: 'auto', width: surfaceWidth }}>
-          <p>Simple Brush</p>
-          <Surface
-            width={surfaceWidth}
-            height={200}
-            viewBox={{
-              x: 0,
-              y: 0,
-              width: surfaceWidth,
-              height: 200,
-            }}
-          >
-            <Brush
-              startIndex={simple.startIndex}
-              endIndex={simple.endIndex}
-              x={100}
-              y={50}
-              width={400}
-              height={40}
-              data={args.data}
-              onChange={handleChange}
-              traveller={renderTraveller}
-              {...args}
-            />
-          </Surface>
-          <p>Brush has specified gap</p>
-          <Surface width={surfaceWidth} height={200}>
-            <Brush
-              startIndex={gap.startIndex}
-              endIndex={gap.endIndex}
-              x={100}
-              y={50}
-              width={400}
-              height={40}
-              data={args.data}
-              gap={5}
-              onChange={handleGapChange}
-              {...args}
-            />
-          </Surface>
-        </div>
-      </ResponsiveContainer>
-    );
-  },
-  args: {
-    ...getStoryArgsFromArgsTypesObject(GeneralProps),
-    data: dateData,
-  },
-  parameters: {
-    controls: { include: Object.keys(GeneralProps) },
-  },
-};
-
-export const PlayBrushMove = {
+export const API = {
   render: (args: Record<string, any>) => {
     return (
       <ResponsiveContainer width="100%" height={400}>
-        <LineChart {...args}>
+        <ComposedChart data={pageData}>
           <Line dataKey="uv" />
+
           <Brush {...args} />
-        </LineChart>
+        </ComposedChart>
       </ResponsiveContainer>
     );
   },
   args: {
     ...getStoryArgsFromArgsTypesObject(GeneralProps),
-    data: numberData,
-  },
-  parameters: {
-    controls: { include: Object.keys(GeneralProps) },
   },
   play: async ({ canvasElement }: { canvasElement: HTMLElement }) => {
     setTimeout(() => {
